backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
fastapi-mail 